make tests discoverable
